[improvment] messing around with syscalls again; fd->iofunc instead of static handling of 0 1 and 2

+++ include/exec/syscall.h
@@ -11,13 +11,13 @@ typedef enum SYSCALL_ERROR_CODE {
 } SYSCALL_ERROR_CODE;
 
 extern SYSCALL_ERROR_CODE handle_syscall(
-    uint64_t syscall_number, 
-    uint64_t rdi,
-    uint64_t rsi,
-    uint64_t rdx,
-    uint64_t r10,
-    uint64_t r8,
-    uint64_t r9,
-    uint64_t user_id,
-    uint64_t process_id
+    uint32_t syscall_number, 
+    uint32_t rdi,
+    uint32_t rsi,
+    uint32_t rdx,
+    uint32_t r10,
+    uint32_t r8,
+    uint32_t r9,
+    uint32_t user_id,
+    uint32_t process_id
 );
